Added a simple menu including a logout button to navbar.
